extract gameUI
