+++ src/html/index.html
@@ -26,7 +26,7 @@
       <div class="card d-flex flex-column p-3 overlay">
         <div class="d-flex flex-row align-items-center justify-content-start my-2">
           <i class="fa-solid fa-gauge-high me-2"></i>
-          <input type="range" class="speed-range" id="speed" name="speed" onchange="window.app.onSpeedChange()" min="0" max="3000">
+          <input type="range" class="speed-range" id="speed" name="speed" onchange="window.app.gameUI.onSpeedChange()" min="0" max="3000">
         </div>
         <div class="d-flex flex-row align-items-center justify-content-start my-2">
           <i class="fa-solid fa-globe me-2"></i>
@@ -37,12 +37,12 @@
     </div>
     <div class="d-flex flex-column justify-content-end align-items-center mb-3 h-100">
       <div class="d-flex flex-row flex-wrap justify-content-center mt-3 w-100">
-        <button type="button" class="btn btn-main btn-success text-uppercase rounded-pill text-nowrap m-2" onclick="window.app.onStart()">Start</button>
-        <button type="button" class="btn btn-main btn-danger text-uppercase rounded-pill text-nowrap m-2" onclick="window.app.onStop()">Stop</button>
-        <button type="button" class="btn btn-main btn-primary text-uppercase rounded-pill text-nowrap m-2" onclick="window.app.onBack()">« Back</button>
-        <button type="button" class="btn btn-main btn-primary text-uppercase rounded-pill text-nowrap m-2" onclick="window.app.onNext()">Next »</button>
-        <button type="button" class="btn btn-main btn-danger text-uppercase rounded-pill text-nowrap m-2" onclick="window.app.onReset()">Reset</button>
-        <button type="button" class="btn btn-main btn-danger text-uppercase rounded-pill text-nowrap m-2" onclick="window.app.onClear()">Clear</button>
+        <button type="button" class="btn btn-main btn-success text-uppercase rounded-pill text-nowrap m-2" onclick="window.app.gameUI.onStart()">Start</button>
+        <button type="button" class="btn btn-main btn-danger text-uppercase rounded-pill text-nowrap m-2" onclick="window.app.gameUI.onStop()">Stop</button>
+        <button type="button" class="btn btn-main btn-primary text-uppercase rounded-pill text-nowrap m-2" onclick="window.app.gameUI.onBack()">« Back</button>
+        <button type="button" class="btn btn-main btn-primary text-uppercase rounded-pill text-nowrap m-2" onclick="window.app.gameUI.onNext()">Next »</button>
+        <button type="button" class="btn btn-main btn-danger text-uppercase rounded-pill text-nowrap m-2" onclick="window.app.gameUI.onReset()">Reset</button>
+        <button type="button" class="btn btn-main btn-danger text-uppercase rounded-pill text-nowrap m-2" onclick="window.app.gameUI.onClear()">Clear</button>
       </div>
 
       <div class="d-flex flex-row flex-wrap justify-content-center align-items-center mt-3 w-100">
